requirements.txt
asgiref==3.8.1
Django==5.2.3
django-filter==24.1
djangorestframework==3.16.0
drf-spectacular==0.27.1
Pillow==10.2.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
sqlparse==0.5.3
